feat: v2 commit

+++ contracts/interfaces/IPairCallee.sol
@@ -1,5 +1,10 @@
 pragma solidity 0.8.13;
 
 interface IPairCallee {
-    function hook(address sender, uint amount0, uint amount1, bytes calldata data) external;
+    function hook(
+        address sender,
+        uint256 amount0,
+        uint256 amount1,
+        bytes calldata data
+    ) external;
 }
