+++ contracts/interfaces/IVeArtProxy.sol
@@ -1,5 +1,10 @@
 pragma solidity 0.8.13;
 
 interface IVeArtProxy {
-    function _tokenURI(uint _tokenId, uint _balanceOf, uint _locked_end, uint _value) external pure returns (string memory output);
+    function _tokenURI(
+        uint256 _tokenId,
+        uint256 _balanceOf,
+        uint256 _locked_end,
+        uint256 _value
+    ) external pure returns (string memory output);
 }
